dubbo服务遗漏的代码

+++ user-thrift-service-api/thrift/user_service.thrift
@@ -1,7 +1,7 @@
 namespace java com.xuehui.thrift.user
 
 struct UserInfo{
-    1:i32 id,
+    1:i64 id,
     2:string username,
     3:string password,
     4:string realName,
